feat: password reset functionality
generate reset token for user with posted email
